cloudfunctions/load_csv_bigquery/requirements.txt
functions-framework
cloudevents
google-cloud-storage
google-cloud-bigquery